refactor tests of binomial and cauchy
